+++ crates/wit-encoder/src/ty.rs
@@ -25,7 +25,7 @@ pub enum Type {
     Borrow(Ident),
     Option(Box<Type>),
     Result(Box<Result_>),
-    List(Box<Type>, Option<usize>),
+    List(Box<Type>, Option<u32>),
     Tuple(Tuple),
     Future(Option<Box<Type>>),
     Stream(Option<Box<Type>>),
@@ -55,7 +55,7 @@ impl Type {
     pub fn result_empty() -> Self {
         Type::Result(Box::new(Result_::empty()))
     }
-    pub fn list(type_: Type, size: Option<usize>) -> Self {
+    pub fn list(type_: Type, size: Option<u32>) -> Self {
         Type::List(Box::new(type_), size)
     }
     pub fn tuple(types: impl IntoIterator<Item = Type>) -> Self {
